Update StatsView xib for first viewcontroller
