+++ PokemonApp/Controller/ViewControllerDetalles.swift
@@ -9,30 +9,13 @@ class ViewControllerDetalles: UIViewController {
     @IBOutlet weak var elementIcon: UIImageView!
     @IBOutlet weak var viewElement: UIView!
     @IBOutlet weak var descriptionLabel: UILabel!
-    @IBOutlet weak var stat1: UILabel!
-    @IBOutlet weak var stat2: UILabel!
-    @IBOutlet weak var stat3: UILabel!
-    @IBOutlet weak var stat4: UILabel!
-    @IBOutlet weak var stat5: UILabel!
-    @IBOutlet weak var stat6: UILabel!
-    
-    @IBOutlet weak var LabelStats1: UILabel!
-    @IBOutlet weak var LabelStats2: UILabel!
-    @IBOutlet weak var LabelStats3: UILabel!
-    @IBOutlet weak var LabelStats4: UILabel!
-    @IBOutlet weak var LabelStats5: UILabel!
-    @IBOutlet weak var LabelStats6: UILabel!
-    
-    @IBOutlet weak var progressStat1: UIProgressView!
-    @IBOutlet weak var progressStat2: UIProgressView!
-    @IBOutlet weak var progressStat3: UIProgressView!
-    @IBOutlet weak var progressStat4: UIProgressView!
-    @IBOutlet weak var progressStat5: UIProgressView!
-    @IBOutlet weak var progressStat6: UIProgressView!
+
     
     @IBOutlet weak var statsButton: UIButton!
     @IBOutlet weak var evolutionButton: UIButton!
     @IBOutlet weak var movesButton: UIButton!
+    
+    @IBOutlet weak var statsView: StatsView!
     
     var selectedPokemon: String = ""
     var selectedPokemonImage: UIImage?
@@ -44,7 +27,8 @@ class ViewControllerDetalles: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         setupButtons()
-        loaditems()
+        
+       // loaditems()
     }
     
     override func viewWillAppear(_ animated: Bool) {
@@ -91,43 +75,7 @@ class ViewControllerDetalles: UIViewController {
         statsButton.backgroundColor = #colorLiteral(red: 0.3951376379, green: 0.6996766925, blue: 0.8580685258, alpha: 1)
         statsButton.setTitleColor(.white, for: .normal)
     }
-    
-    private func loaditems() {
-        renameLabelStats()
         
-        if let stats = stats {
-            assignmentValueProgressBar(stats)
-            septupLabelProgressBar(stats)
-        }
-    }
-    
-    private func assignmentValueProgressBar(_ stats: [Float]) {
-        progressStat1.progress = Float(stats[0])
-        progressStat2.progress = Float(stats[1])
-        progressStat3.progress = Float(stats[2])
-        progressStat4.progress = Float(stats[3])
-        progressStat5.progress = Float(stats[4])
-        progressStat6.progress = Float(stats[5])
-    }
-    
-    private func septupLabelProgressBar(_ stats: [Float]) {
-        stat1.text = String(Int(stats[0] * 100))
-        stat2.text = String(Int(stats[1] * 100))
-        stat3.text = String(Int(stats[2] * 100))
-        stat4.text = String(Int(stats[3] * 100))
-        stat5.text = String(Int(stats[4] * 100))
-        stat6.text = String(Int(stats[5] * 100))
-    }
-    
-    private func renameLabelStats() {
-        LabelStats1.text = "Hp"
-        LabelStats2.text = "Attack"
-        LabelStats3.text = "Defense"
-        LabelStats4.text = "Sp. Atk"
-        LabelStats5.text = "Sp. Def"
-        LabelStats6.text = "Speed"
-    }
-    
     private func setupElement() {
         elementLabel.text = selectedPokemonTextElement?.capitalized
         elementIcon.image = selectedPokemonIconoElement
